+++ examples/angular/src/polyfills.ts
@@ -1,8 +1,3 @@
-(window as any).global = window;
-(window as any).process = {
-    env: { DEBUG: undefined },
-};
-
 /*
  * Zone JS is required by default for Angular itself.
  */