+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added `doesBroadcastChangestream()` to `RxStorageStatics`
 - Fix (replication primitives) only drop pulled documents when a relevant document was changed locally.
 - Add `_meta` property to stored document data.
+- Automatically polyfill pouchdb env vars on first call to `getRxStoragePouch()`.
 
 
 ### 11.6.0 (4 February 2022)